requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
flask==3.0.3
opencv-python==4.10.0.84
numpy==2.1.1
pyserial==3.5
pyautogui==0.9.54
pynput==1.7.7
